refactor: move protocol types to src/protocol/types.ts
